requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.16
jinja2>=3.1.0
python-multipart>=0.0.9
